requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
rdkit-pypi